test: relocate coverage tests to follow Go file naming convention
Move tests added in 34d1e699 into files matching the source file under
test, per Go convention (xxx.go -> xxx_test.go):

- internal/roots/roots_coverage_test.go -> merged into roots_test.go
  (tests ClientSupportsRoots in roots.go).
- internal/tools/elicitationtools/coverage_test.go -> split into:
    - elicitationtools_test.go for the four FinalConfirmCancel tests
      (cover IssueCreate/MRCreate/ReleaseCreate/ProjectCreate in
      elicitationtools.go).
    - register_test.go (new) for the four round-trip cancellation
      tests covering the closures in register.go.

No test logic changes; only file relocations.
